Renamed utils as common
